requirements.txt
Flask
Flask-Session
psycopg2-binary
SQLAlchemy


import os
import psycopg2, requests, json
from tables import TableMake

from flask import Flask, session, render_template, request, redirect, url_for, jsonify, flash
from flask_session import Session
from sqlalchemy import create_engine, MetaData, Table, Column, String, select
from sqlalchemy.orm import scoped_session, sessionmaker
